signin w/ try except
