+++ env-example
@@ -1,0 +1,2 @@
+FOGO_CRUZADO_EMAIL=usuario@host.com
+FOGO_CRUZADO_PASSWORD=password
